backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
bcrypt==4.0.1
pyjwt==2.8.0
werkzeug==2.3.7
numpy==1.24.3
opencv-python==4.8.1.78
mediapipe==0.10.0
scikit-learn==1.3.0
joblib==1.3.2
python-dotenv==1.0.0